fix variable lookup
